+++ src/constants/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    COSMOS_DIRECTORY_URL: "https://chains.cosmos.directory/",
+}